move xarray functions to xrcompat
